+++ sse/client.go
